Improve docker stack
PHPMyAdmin has been added to the stack to easily browse the database.
It is accessible via http://localhost:8080, and will only talk to the
mariadb container.

All containers now have a name defined, so we can more easily reference them.

MariaDB now requires a password, and we move to the simple MYSQL_DATABASE
environment variable to create the default database.

The wrapper script now has a "clean" option, which will bring the containers
down and clean up volumes.
We give all the containers a name